requirements.txt
pandas
google-genai
flask
requests
jsonify
streamlit